Updating the url.

+++ src/SideEffects.js
@@ -12,29 +12,43 @@ import * as Keys from "./model/keys";
 const scope = "se.kth.dinnerplanner.siddaniel.";
 
 const types = {
-    fetchDish: scope + "fetchDish"
+    fetchDish: scope + "fetchDish",
+    updateUrl: scope + "updateUrl"
 };
 
 function fetchDish(dishId) {
-    function success(dish) {
-        return Actions.fetchedDish(dish);
-    }
-
-    function error(dishId, reason) {
-        return Actions.failedToFetchDish({dishId: dishId, reason: reason});
-    }
-
     return {
         type: types.fetchDish,
         key: dishId,
-        successAction: success,
-        errorAction: error,
+        successAction: (dish) => Actions.fetchedDish(dish),
+        errorAction: (dishId, reason) => Actions.failedToFetchDish({dishId: dishId, reason: reason}),
+    }
+}
+
+function updateUrl(url) {
+    return {
+        type: types.updateUrl,
+        key: url.toString(),
     }
 }
 
 export function renderSideEffects(state) {
+    let url = new URL(window.location);
+    url.search = "";
+    url.pathname = `/${state.get('page')}`;
+    const selectedDishId = Core.getSelectedDishId(state);
+    if (selectedDishId) {
+        url.searchParams.set('id', selectedDishId);
+    }
     return Set()
-        .add(Core.getSelectedDishId(state) && !Core.getFullDataOnSelectedDish(state) && fetchDish(getSelectedDishId(state)))
+        .add(
+            Core.getSelectedDishId(state)
+            && !Core.getFullDataOnSelectedDish(state)
+            && fetchDish(getSelectedDishId(state))
+        )
+        .add(
+            updateUrl(url)
+        )
         .filter(x => x);
 }
 
@@ -43,17 +57,22 @@ function id(sideEffect) {
 }
 
 export function sideEffectMapper(sideEffects, dispatch) {
-    let pendings = {};
     sideEffects.forEach(sideEffect => {
         switch (sideEffect.type) {
             case types.fetchDish:
                 let promise = Services.fetchDish(sideEffect.key);
                 promise.then(dish => dispatch(sideEffect.successAction(dish)), (error) => dispatch(sideEffect.errorAction(sideEffect.key, error)));
-                pendings[id(sideEffect)] = {};
                 break;
+            case types.updateUrl:
+                let url = new URL(sideEffect.key);
+                history.pushState(
+                    {todo:"back navigation not implemented"},
+                    url.pathname,
+                    `${url.pathname}${url.search}`
+                )
         }
     });
-    return {reduxAction: {type: "sideEffectCompleted"}, pending: pendings};
+    return {reduxAction: {type: "sideEffectCompleted"}, pending: {}};
 }
 
 function splitSideEffects(renderedSideEffects, pendingSideEffects) {
@@ -72,12 +91,18 @@ export class SideEffector {
     constructor(mapper, dispatch) {
         this._mapper = mapper;
         this._pending = Map();
+        this._lastEffectIdsToPerform = Set();
         this._dispatch = dispatch;
     }
 
     perform(sideEffects) {
         const pendingIn = this._pending;
         const {pendingToCancel, effectsToPerform} = splitSideEffects(sideEffects, this._pending);
+        const effectIdsToPerform = Set(effectsToPerform.map(id));
+        if (this._lastEffectIdsToPerform.equals(effectIdsToPerform)) {
+            return
+        }
+        this._lastEffectIdsToPerform = effectIdsToPerform;
         pendingToCancel.valueSeq().forEach(pending => pending.cancel && pending.cancel());
         const pendingSubCancelled = pendingIn.deleteAll(pendingToCancel.keys());
         console.log("effects to perform");
